Functioning images
geosupport-docker:<version>-dist is based on scratch image and contains
only the repackaged Geosupport distribution and volume declaration.

geosupport-docker:<version> is similar to the previous version except
smaller in size and is built from more repeatable generated sources. It
also uses COPY to copy the tgz from the geosupport-docker:<version>-dist
image. This allows for using different parent images when running the
generated.sh script.

Use scratch image to reduce image size of 'distributor' stage from ~400
mb to ~200 mb. The 'hello' executable added with this commit will be
modified or deleted if this approach continues. It was taken from
Docker's example 'hello-world' image.

General cleanup and documentation are required before this can be
considered a release candidate.