Functioning images for next release
geosupport-docker:<version>-dist is based on scratch image and contains
only the repackaged Geosupport distribution and volume declaration.

geosupport-docker:<version> is similar to the previous version except
smaller in size and is built from more repeatable generated sources. It
also uses COPY to copy the tgz from the geosupport-docker:<version>-dist
image. This allows for using different parent images when running the
generated.sh script.

General cleanup and documentation are required before this can be
considered a release candidate.